third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-container-only-white-space-expected.txt

This is an element
This is an element
PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test tests that a anchor element with only white spaces gives the correct size in ContainerNode::getRect().
